♻️ refactor: split config.go into focused files
Break 433-line config.go into 5 files by responsibility:
- config.go: core structs, Load/LoadFrom
- channels.go: channel configs + boolDefault helper
- model.go: model types, tier constants, resolution
- paths.go: AnnaHome (cached via sync.Once), path helpers
- state.go: state persistence (SaveModelSelection, applyState)

Deduplicate 8 copies of *bool nil-check into boolDefault(),
cache AnnaHome() to avoid repeated syscalls, and collapse
duplicate ResolveModelID default case.
